requirements.txt
Pillow
PyQt5
numpy
opencv-python
scikit-image
scipy
shapely
svgpathtools
